create StatusPill and FitlerPill
