+++ frontend/packages/client/src/App.sass
@@ -879,7 +879,16 @@ input[type=range].slider::-ms-fill-upper
 .search-input
 	width: 480px		
 .pill-border
-	border: 2px solid #000000
+	border: 2px solid
+
+.has-border-success
+	border-color: $success
+.has-border-danger
+	border-color: $danger
+.has-border-yellow
+	border-color: $yellow
+.has-border-orange
+	border-color: $orange
 
 @include mobile	
 	.search-container
