+++ frontend/packages/client/src/components/index.js
@@ -46,5 +46,6 @@ export { default as Input } from './Input';
 export { default as TooltipMessage } from './TooltipMessage';
 export { default as FadeInOut } from './FadeInOut';
 export { default as ErrorModal } from './ErrorModal';
-export { default as Pill } from './Pill';
+export { default as StatusPill } from './StatusPill';
+export { default as FilterPill } from './FilterPill';
 export * from './Proposal';
